Use library access code if loading without item
