+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/browsing/library/access/DelegatingLibraryProvider.kt
@@ -4,7 +4,7 @@ import com.lasthopesoftware.bluewater.client.browsing.library.repository.Library
 import com.lasthopesoftware.bluewater.shared.policies.ApplyExecutionPolicies
 
 class DelegatingLibraryProvider(inner: ILibraryProvider, policies: ApplyExecutionPolicies) : ILibraryProvider by inner {
-	private val getLibraryFunc = policies.applyPolicy(inner::getLibrary)
+	private val getLibraryFunc = policies.applyPolicy(inner::promiseLibrary)
 
-	override fun getLibrary(libraryId: LibraryId) = getLibraryFunc(libraryId)
+	override fun promiseLibrary(libraryId: LibraryId) = getLibraryFunc(libraryId)
 }
