+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/stored/library/SyncLibraryProvider.kt
@@ -9,8 +9,8 @@ class SyncLibraryProvider(private val libraryProvider: ILibraryProvider) : ILibr
 	override val allLibraries: Promise<Collection<Library>>
 		get() = libraryProvider.allLibraries.then { l -> l.map { it.transformConnectionSetting() } }
 
-	override fun getLibrary(libraryId: LibraryId): Promise<Library?> =
-		libraryProvider.getLibrary(libraryId).then { it?.transformConnectionSetting() }
+	override fun promiseLibrary(libraryId: LibraryId): Promise<Library?> =
+		libraryProvider.promiseLibrary(libraryId).then { it?.transformConnectionSetting() }
 
 	private fun Library.transformConnectionSetting() = setLocalOnly(isSyncLocalConnectionsOnly)
 }
