+++ projectBlueWater/src/test/java/com/lasthopesoftware/bluewater/client/browsing/items/list/GivenALibrary/AndAnItem/AndItHasChildItems/WhenPlayingAShuffledChildItem.kt
@@ -18,28 +18,30 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 
-private const val libraryId = 773
-private const val itemId = 107
-private const val itemValue = "height"
+private const val libraryId = 411
+private const val itemId = 883
+private const val itemValue = "can"
 
 private var playedFileList = ""
 
-class WhenPlayingAChildItem {
+class WhenPlayingAShuffledChildItem {
+
 	private val viewModel by lazy {
 		val itemProvider = mockk<ProvideItems>().apply {
 			every { promiseItems(LibraryId(libraryId), ItemId(itemId)) } returns listOf(
-				Item(611),
-				Item(306),
-				Item(867),
-				Item(623),
-				Item(335),
-				Item(983),
+				Item(495),
+				Item(74),
+				Item(525),
+				Item(787),
+				Item(460),
+				Item(107),
+				Item(923),
 			).toPromise()
 		}
 
 		val itemStringListProvider = mockk<ProvideFileStringListForItem>().apply {
-			every { promiseFileStringList(LibraryId(libraryId), ItemId(867), FileListParameters.Options.None) } returns Promise(
-				"2;-1;959;191;559;815;165;"
+			every { promiseFileStringList(LibraryId(libraryId), ItemId(460), FileListParameters.Options.Shuffled) } returns Promise(
+				"2;-1;920;388;906;356204;641;221;889;"
 			)
 		}
 
@@ -52,7 +54,8 @@ class WhenPlayingAChildItem {
 		ItemListViewModel(
             itemProvider,
             mockk(relaxed = true, relaxUnitFun = true),
-            FakeStoredItemAccess(),
+			mockk(),
+			FakeStoredItemAccess(),
             itemStringListProvider,
             controlNowPlaying,
             mockk(),
@@ -63,11 +66,11 @@ class WhenPlayingAChildItem {
 	@BeforeAll
 	fun act() {
 		viewModel.loadItem(LibraryId(libraryId), Item(itemId, itemValue)).toExpiringFuture().get()
-		viewModel.items.value[2].play().toExpiringFuture().get()
+		viewModel.items.value[4].playShuffled().toExpiringFuture().get()
 	}
 
 	@Test
 	fun `then the child items are played`() {
-		assertThat(playedFileList).isEqualTo("2;-1;959;191;559;815;165;")
+		assertThat(playedFileList).isEqualTo("2;-1;920;388;906;356204;641;221;889;")
 	}
 }
