+++ projectBlueWater/src/test/java/com/lasthopesoftware/bluewater/client/browsing/items/list/GivenALibrary/AndAnItem/AndItHasChildItems/WhenSyncingAChildItem.kt
@@ -33,6 +33,7 @@ class WhenSyncingAChildItem {
 		ItemListViewModel(
 			itemProvider,
 			mockk(relaxed = true, relaxUnitFun = true),
+			mockk(),
 			storedItemAccess,
 			mockk(),
 			mockk(),
